Refactored ArtistPage to Use Styled Components
